requirements.txt
asn1crypto==0.24.0
Authlib==0.10
awscli==1.16.65
boto3==1.9.38
botocore==1.12.55
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.4.2
dnslib==0.9.7
docutils==0.14
fierce==1.2.2
Flask==1.0.2
future==0.17.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
pyasn1==0.4.4
pycparser==2.19
python-dateutil==2.7.5
python-whois==0.7.0
PyYAML==3.13
requests==2.20.1
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
Flask-Menu==0.7
tldextract==2.2.0
